refactor:change to brast class
