+++ frontend/src/types/GameObjectFactory.d.ts
@@ -16,7 +16,7 @@ declare global {
       ) => MyPlayer;
 
       bomb: (x: number, y: number, bombStrength: number) => Bomb;
-      bombExplosion: (x: number, y: number, bombStrength: number) => BombExplosion;
+      blast: (x: number, y: number, bombStrength: number) => BombExplosion;
     }
   }
 }
